recovered/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
requests==2.31.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.24.0
aiofiles>=23.0.0
python-multipart>=0.0.5
jinja2>=3.0.0
